+++ example/android/app/src/main/kotlin/com/example/notifications_handler_demo/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.notifications_handler_demo
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}