+++ example/lib/main.dart
@@ -1,76 +1,55 @@
+import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_notifications_handler/firebase_notifications_handler.dart';
 import 'package:flutter/material.dart';
+import 'package:notifications_handler_demo/firebase_options.dart';
+import 'package:notifications_handler_demo/screens/splash_screen.dart';
+import 'package:notifications_handler_demo/utils/app_theme.dart';
+import 'package:notifications_handler_demo/utils/globals.dart';
+import 'package:notifications_handler_demo/utils/helpers.dart';
+import 'package:notifications_handler_demo/utils/route_generator.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  // TODO: add firebase_core and call
-  // await Firebase.initializeApp();
-  runApp(_MainApp());
+  await Firebase.initializeApp(options: DefaultFirebaseOptions.currentPlatform);
+  runApp(const _MainApp());
 }
-
-String? fcmToken;
 
 class _MainApp extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      navigatorKey: FirebaseNotificationsHandler.navigatorKey,
-      home: const HomeScreen(),
-    );
-  }
-}
+  static const id = '_MainApp';
 
-class HomeScreen extends StatelessWidget {
-  const HomeScreen({
+  const _MainApp({
     Key? key,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return FirebaseNotificationsHandler(
+      defaultNavigatorKey: Globals.navigatorKey,
       onOpenNotificationArrive: (_, payload) {
-        debugPrint(
-          "Notification received while app is open with payload $payload",
+        log(
+          id,
+          msg: "Notification received while app is open with payload $payload",
         );
       },
       onTap: (navigatorState, appState, payload) {
-        debugPrint("Notification tapped with $appState & payload $payload");
-
-        navigatorState.currentState!.pushNamed('newRouteName');
-        // OR
-        final context = navigatorState.currentContext!;
-        Navigator.pushNamed(context, 'newRouteName');
+        showSnackBar('appState: $appState\npayload: $payload');
+        log(
+          id,
+          msg: "Notification tapped with $appState & payload $payload",
+        );
       },
-      onFCMTokenInitialize: (_, token) => fcmToken = token,
-      onFCMTokenUpdate: (_, token) {
-        fcmToken = token;
-        // await User.updateFCM(token);
-      },
-      child: SafeArea(
-        child: Scaffold(
-          body: Center(
-            child: Text(
-              'HomeScreen',
-              style: Theme.of(context).textTheme.headline3,
-            ),
-          ),
-        ),
+      onFCMTokenInitialize: (_, token) => Globals.fcmToken = token,
+      onFCMTokenUpdate: (_, token) => Globals.fcmToken = token,
+      child: MaterialApp(
+        debugShowCheckedModeBanner: false,
+        title: 'FirebaseNotificationsHandler Demo',
+        navigatorKey: Globals.navigatorKey,
+        scaffoldMessengerKey: Globals.scaffoldMessengerKey,
+        theme: AppTheme.lightTheme,
+        darkTheme: AppTheme.darkTheme,
+        onGenerateRoute: RouteGenerator.generateRoute,
+        initialRoute: SplashScreen.id,
       ),
     );
   }
-}
-
-void sendExampleNotification() async {
-  await FirebaseNotificationsHandler.sendNotification(
-    cloudMessagingServerKey: '<YOUR_CLOUD_MESSAGING_SERVER_KEY>',
-    title: 'This is a test notification',
-    body: 'This describes this notification',
-    fcmTokens: [
-      'fcmDeviceToken1',
-      'fcmDeviceToken2',
-    ],
-    payload: {
-      'key': 'value',
-    },
-  );
 }