+++ example/lib/screens/splash_screen.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+import 'package:notifications_handler_demo/screens/home_screen.dart';
+import 'package:notifications_handler_demo/widgets/custom_loader.dart';
+
+class SplashScreen extends StatefulWidget {
+  static const id = 'SplashScreen';
+
+  const SplashScreen({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  _SplashScreenState createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> {
+  @override
+  void initState() {
+    (() async {
+      await Future.delayed(Duration.zero);
+      Navigator.pushReplacementNamed(
+        context,
+        HomeScreen.id,
+      );
+    })();
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return const SafeArea(
+      child: Scaffold(
+        body: CustomLoader(),
+      ),
+    );
+  }
+}