+++ example/lib/utils/route_generator.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+import 'package:notifications_handler_demo/screens/home_screen.dart';
+import 'package:notifications_handler_demo/screens/splash_screen.dart';
+import 'package:notifications_handler_demo/utils/helpers.dart';
+
+class RouteGenerator {
+  static const _id = 'RouteGenerator';
+
+  static Route<dynamic> generateRoute(RouteSettings settings) {
+    final args = settings.arguments as dynamic;
+    log(_id, msg: "Pushed ${settings.name}(${args ?? ''})");
+    switch (settings.name) {
+      case SplashScreen.id:
+        return _route(const SplashScreen());
+      case HomeScreen.id:
+        return _route(const HomeScreen());
+      default:
+        return _errorRoute(settings.name);
+    }
+  }
+
+  static MaterialPageRoute _route(Widget widget) =>
+      MaterialPageRoute(builder: (context) => widget);
+
+  static Route<dynamic> _errorRoute(String? name) {
+    return MaterialPageRoute(
+      builder: (_) => Scaffold(
+        body: Center(
+          child: Text('ROUTE \n\n$name\n\nNOT FOUND'),
+        ),
+      ),
+    );
+  }
+}